feat: unify integration tests under one file with different test cases for NoisyNet and RocksDB storage layer.
