+++ test/scripts/bspNetBootstrap.ts
@@ -1,16 +1,20 @@
 import { setTimeout } from "node:timers/promises";
 import {
-  DUMMY_MSP_ID,
-  NODE_INFOS,
   createApiObject,
   createCheckBucket,
+  DUMMY_MSP_ID,
   getContainerPeerId,
+  NODE_INFOS,
+  registerToxics,
   runBspNet,
   shUser,
-  type BspNetApi
+  type BspNetApi,
+  type ToxicInfo
 } from "../util";
 
 let api: BspNetApi | undefined;
+const NOISY = process.env.NOISY === "1" ?? false;
+const ROCKSDB = process.env.ROCKSDB === "1" ?? false;
 
 const CONFIG = {
   bucketName: "nothingmuch-0",
@@ -19,7 +23,39 @@ const CONFIG = {
 };
 
 async function bootStrapNetwork() {
-  await runBspNet();
+  await runBspNet(NOISY, ROCKSDB);
+
+  if (NOISY) {
+    // For more info on the kind of toxics you can register,
+    // see: https://github.com/Shopify/toxiproxy?tab=readme-ov-file#toxics
+    const reqToxics = [
+      {
+        type: "latency",
+        name: "lag-down",
+        stream: "upstream",
+        toxicity: 0.8,
+        attributes: {
+          latency: 25,
+          jitter: 7
+        }
+      },
+      {
+        type: "bandwidth",
+        name: "low-band",
+        // Setting as upstream simulates slow user connection
+        stream: "upstream",
+        // 50% of the time, the toxic will be applied
+        toxicity: 0.5,
+        attributes: {
+          // 10kbps
+          rate: 10
+        }
+      }
+    ] satisfies ToxicInfo[];
+
+    await registerToxics(reqToxics);
+  }
+
   api = await createApiObject(`ws://127.0.0.1:${NODE_INFOS.user.port}`);
 
   const newBucketEventDataBlob = await createCheckBucket(api, CONFIG.bucketName);
@@ -54,7 +90,12 @@ async function bootStrapNetwork() {
   // Seal the block from BSP volunteer
   await setTimeout(1000);
   await api.sealBlock();
-  console.log("✅ BSPNet Bootstrap success");
+
+  if (NOISY) {
+    console.log("✅ NoisyNet Bootstrap success");
+  } else {
+    console.log("✅ BSPNet Bootstrap success");
+  }
 }
 
 bootStrapNetwork()
